document_portal/requirements.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_google_genai
langchain_community
pypdf
faiss-cpu
structlog
PyMuPDF
fitz
pandas
langchain-core[tracing]
streamlit
pytest
-e .